+++ basketserver/src/main/java/de/upb/crypto/incentive/basketserver/BasketserverApplication.java
@@ -1,0 +1,13 @@
+package de.upb.crypto.incentive.basketserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BasketserverApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BasketserverApplication.class, args);
+    }
+
+}