+++ basketserver/src/main/java/de/upb/crypto/incentive/basketserver/model/BasketItem.java
@@ -1,0 +1,16 @@
+package de.upb.crypto.incentive.basketserver.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * Utility dataclass for representing Items in a Basket.
+ */
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class BasketItem {
+    Item item;
+    int count;
+}