+++ basketserver/src/main/java/de/upb/crypto/incentive/basketserver/model/requests/PayBasketRequest.java
@@ -1,0 +1,18 @@
+package de.upb.crypto.incentive.basketserver.model.requests;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+/**
+ * Dataclass for pay basket request body.
+ */
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class PayBasketRequest {
+    UUID basketId;
+    long value;
+}