+++ basketserver/src/test/java/de/upb/crypto/incentive/basketserver/BasketserverApplicationTests.java
@@ -1,0 +1,13 @@
+package de.upb.crypto.incentive.basketserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BasketserverApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}